+++ src/tedi/components/base/text/text.component.ts
@@ -5,13 +5,12 @@ import {
   input,
 } from "@angular/core";
 
+export type HeadingLevel = 1 | 2 | 3 | 4 | 5 | 6;
+
+export type HeadingModifiers = `h${HeadingLevel}`;
+
 export type TextModifiers =
-  | "h1"
-  | "h2"
-  | "h3"
-  | "h4"
-  | "h5"
-  | "h6"
+  | HeadingModifiers
   | "normal"
   | "small"
   | "extra-small"